+++ include/ObscureEngine/WS/Window.h
@@ -21,7 +21,7 @@ namespace ObscureEngine
       void swapBuffers();
       void makeContextCurrent();
 
-      void clear(uint8_t red, uint8_t green, uint8_t blue, uint8_t alpha);
+      void clear(float red, float green, float blue, float alpha);
 
       void pollEvent();
       void close();
